feat: move pages to /app
- /pilotage
- /mes-jeunes
- /profil
- /reaffectation
- /recherche-offres
- /raccourci
- /
