+++ app/(connected)/(with-sidebar)/(with-chat)/(index)/HomePage.tsx
@@ -1,20 +1,13 @@
+'use client'
+
 import { withTransaction } from '@elastic/apm-rum-react'
-import { GetServerSideProps, GetServerSidePropsResult } from 'next'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
 
 import RenseignementAgenceModal from 'components/RenseignementAgenceModal'
 import RenseignementEmailModal from 'components/RenseignementEmailModal'
 import RenseignementStructureModal from 'components/RenseignementStructureModal'
-import {
-  aEtablissement,
-  Conseiller,
-  doitSignerLesCGU,
-  estMilo,
-  estPassEmploi,
-  StructureConseiller,
-} from 'interfaces/conseiller'
-import { PageProps } from 'interfaces/pageProps'
+import { StructureConseiller } from 'interfaces/conseiller'
 import { Agence } from 'interfaces/referentiel'
 import { AlerteParam } from 'referentiel/alerteParam'
 import { useAlerte } from 'utils/alerteContext'
@@ -25,17 +18,16 @@ import {
 } from 'utils/analytics/matomo'
 import useMatomo from 'utils/analytics/useMatomo'
 import { useConseiller } from 'utils/conseiller/conseillerContext'
-import { ApiError } from 'utils/httpClient'
 import { usePortefeuille } from 'utils/portefeuilleContext'
 
-interface HomePageProps extends PageProps {
+type HomePageProps = {
   redirectUrl: string
   afficherModaleAgence: boolean
   afficherModaleEmail: boolean
   referentielAgences?: Agence[]
 }
 
-function Home({
+function HomePage({
   afficherModaleAgence,
   afficherModaleEmail,
   redirectUrl,
@@ -68,7 +60,7 @@ function Home({
   }
 
   async function redirectToUrl() {
-    await router.replace(redirectUrl)
+    router.replace(redirectUrl)
   }
 
   function trackContacterSupport(etablissement: string) {
@@ -124,87 +116,4 @@ function Home({
   )
 }
 
-export const getServerSideProps: GetServerSideProps<HomePageProps> = async (
-  context
-): Promise<GetServerSidePropsResult<HomePageProps>> => {
-  const { default: withMandatorySessionOrRedirect } = await import(
-    'utils/auth/withMandatorySessionOrRedirect'
-  )
-  const sessionOrRedirect = await withMandatorySessionOrRedirect(context)
-  if (!sessionOrRedirect.validSession) {
-    return { redirect: sessionOrRedirect.redirect }
-  }
-
-  const {
-    session: { user, accessToken },
-  } = sessionOrRedirect
-
-  const sourceQueryParam = context.query.source
-    ? `?source=${context.query.source}`
-    : ''
-
-  const { getConseillerServerSide } = await import(
-    'services/conseiller.service'
-  )
-  let conseiller: Conseiller | undefined
-  try {
-    conseiller = await getConseillerServerSide(user, accessToken)
-  } catch (e) {
-    if (e instanceof ApiError && e.statusCode === 401) {
-      return {
-        redirect: {
-          destination: '/api/auth/federated-logout',
-          permanent: false,
-        },
-      }
-    }
-    throw e
-  }
-  if (!conseiller) {
-    throw new Error(`Conseiller ${user.id} inexistant`)
-  }
-
-  if (doitSignerLesCGU(conseiller))
-    return {
-      redirect: {
-        destination: '/consentement-cgu',
-        permanent: false,
-      },
-    }
-
-  const redirectUrl =
-    (context.query.redirectUrl as string) ?? '/mes-jeunes' + sourceQueryParam
-
-  const emailEstManquant = estMilo(conseiller) && !conseiller.email
-  const agenceEstManquante =
-    !estPassEmploi(conseiller) && !aEtablissement(conseiller)
-
-  if (emailEstManquant || agenceEstManquante) {
-    let props: HomePageProps = {
-      afficherModaleAgence: agenceEstManquante,
-      afficherModaleEmail: emailEstManquant,
-      redirectUrl,
-      pageTitle: 'Accueil',
-    }
-
-    if (!estMilo(conseiller) && agenceEstManquante) {
-      const { getAgencesServerSide } = await import(
-        'services/referentiel.service'
-      )
-      props.referentielAgences = await getAgencesServerSide(
-        conseiller.structure,
-        accessToken
-      )
-    }
-    return { props }
-  }
-
-  return {
-    redirect: {
-      destination: redirectUrl,
-      permanent: false,
-    },
-  }
-}
-
-export default withTransaction(Home.name, 'page')(Home)
+export default withTransaction(HomePage.name, 'page')(HomePage)
