+++ app/(connected)/(with-sidebar)/(with-chat)/reaffectation/page.tsx
@@ -1,0 +1,21 @@
+import { Metadata } from 'next'
+import { notFound } from 'next/navigation'
+import React from 'react'
+
+import ReaffectationPage from 'app/(connected)/(with-sidebar)/(with-chat)/reaffectation/ReaffectationPage'
+import { PageHeaderPortal } from 'components/PageNavigationPortals'
+import { getMandatorySessionServerSide } from 'utils/auth/auth'
+
+export const metadata: Metadata = { title: 'Réaffectation' }
+export default async function Reaffectation() {
+  const { user } = await getMandatorySessionServerSide()
+  if (!user.estSuperviseur) notFound()
+
+  return (
+    <>
+      <PageHeaderPortal header='Réaffectation' />
+
+      <ReaffectationPage estSuperviseurPEBRSA={user.estSuperviseurPEBRSA} />
+    </>
+  )
+}
