+++ app/(connected)/(with-sidebar)/(without-chat)/raccourci/RaccourciPage.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import { withTransaction } from '@elastic/apm-rum-react'
+
+import { TutorielRaccourci } from 'components/TutorielRaccourci'
+import useMatomo from 'utils/analytics/useMatomo'
+
+function RaccourciPage() {
+  useMatomo('Tuto raccourci mobile')
+
+  return <TutorielRaccourci />
+}
+
+export default withTransaction(RaccourciPage.name, 'page')(RaccourciPage)
