+++ tests/integration/NotificationsSonores.test.tsx
@@ -7,7 +7,7 @@ import Layout from 'components/layouts/Layout'
 import { unConseiller } from 'fixtures/conseiller'
 import { desItemsJeunes, unJeuneChat } from 'fixtures/jeune'
 import { BaseJeune, JeuneChat, JeuneFromListe } from 'interfaces/jeune'
-import Profil from 'pages/profil'
+import Profil from 'app/(connected)/(with-sidebar)/(with-chat)/profil/ProfilPage'
 import { getJeunesDuConseillerClientSide } from 'services/jeunes.service'
 import { observeConseillerChats, signIn } from 'services/messages.service'
 import renderWithContexts from 'tests/renderWithContexts'
