+++ tests/pages/offres/RechercheServicesCiviques.page.test.tsx
@@ -6,46 +6,55 @@ import {
   within,
 } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
+import { DateTime } from 'luxon'
 import React from 'react'
 
+import RechercheOffresPage from 'app/(connected)/(with-sidebar)/(with-chat)/offres/RechercheOffresPage'
 import { unConseiller } from 'fixtures/conseiller'
-import { listeBaseAlternances, uneBaseAlternance } from 'fixtures/offre'
-import { desLocalites, unDepartement, uneCommune } from 'fixtures/referentiel'
-import { StructureConseiller } from 'interfaces/conseiller'
-import { BaseOffreEmploi } from 'interfaces/offre'
-import { Localite } from 'interfaces/referentiel'
-import RechercheOffres from 'pages/recherche-offres'
 import {
-  getOffreEmploiClientSide,
-  searchAlternances,
-} from 'services/offres-emploi.service'
-import { getCommunesEtDepartements } from 'services/referentiel.service'
+  listeBaseServicesCiviques,
+  uneBaseServiceCivique,
+} from 'fixtures/offre'
+import { desCommunes, desLocalites, uneCommune } from 'fixtures/referentiel'
+import { StructureConseiller } from 'interfaces/conseiller'
+import { BaseServiceCivique } from 'interfaces/offre'
+import { Localite } from 'interfaces/referentiel'
+import { domainesServiceCivique } from 'referentiel/domaines-service-civique'
+import {
+  getCommunes,
+  getCommunesEtDepartements,
+} from 'services/referentiel.service'
+import { searchServicesCiviques } from 'services/services-civiques.service'
 import { getByTextContent } from 'tests/querySelector'
 import renderWithContexts from 'tests/renderWithContexts'
 
 jest.mock('utils/auth/withMandatorySessionOrRedirect')
-jest.mock('services/offres-emploi.service')
 jest.mock('services/referentiel.service')
+jest.mock('services/services-civiques.service')
 
-describe('Page Recherche Alternances', () => {
-  let alternances: BaseOffreEmploi[]
-  let localites: Localite[]
+describe('Page Recherche Offres Service civique', () => {
+  let servicesCiviques: BaseServiceCivique[]
+  let communes: Localite[]
 
   let rendered: RenderResult
+
   describe('quand le conseiller n’est pas PE BRSA', () => {
     beforeEach(async () => {
-      localites = desLocalites()
-      alternances = listeBaseAlternances()
-      ;(searchAlternances as jest.Mock).mockResolvedValue({
-        metadonnees: { nombreTotal: 37, nombrePages: 4 },
-        offres: alternances,
-      })
+      communes = desCommunes()
+      servicesCiviques = listeBaseServicesCiviques()
       ;(getCommunesEtDepartements as jest.Mock).mockResolvedValue(
         desLocalites()
       )
+      ;(getCommunes as jest.Mock).mockResolvedValue(desCommunes())
+      ;(searchServicesCiviques as jest.Mock).mockResolvedValue({
+        metadonnees: { nombreTotal: 37, nombrePages: 4 },
+        offres: servicesCiviques,
+      })
 
-      rendered = renderWithContexts(<RechercheOffres pageTitle='' />, {})
-      await userEvent.click(screen.getByRole('radio', { name: 'Alternance' }))
+      rendered = renderWithContexts(<RechercheOffresPage />, {})
+      await userEvent.click(
+        screen.getByRole('radio', { name: 'Service civique' })
+      )
     })
 
     it('permet de definir des critères de recherche', () => {
@@ -56,98 +65,52 @@ describe('Page Recherche Alternances', () => {
 
       expect(etape2).toBeInTheDocument()
       expect(
-        within(etape2).getByRole('checkbox', {
-          name: 'Recherche avec un numéro d’offre Pôle emploi',
+        within(etape2).getByRole('combobox', {
+          name: 'Localisation Saisissez une ville',
         })
       ).toBeInTheDocument()
-
-      expect(
-        within(etape2).queryByLabelText(/Numéro d’offre/)
-      ).not.toBeInTheDocument()
-      expect(
-        within(etape2).getByLabelText('Mots clés (Métier, code ROME)')
-      ).toHaveAttribute('type', 'text')
-      expect(
-        within(etape2).getByLabelText('Mots clés (Métier, code ROME)')
-      ).not.toHaveAttribute('disabled')
-      expect(
-        within(etape2).getByRole('combobox', {
-          name: 'Lieu de travail Saisissez une ville ou un département',
-        })
-      ).toBeInTheDocument()
-      expect(
-        within(etape2).getByRole('combobox', {
-          name: 'Lieu de travail Saisissez une ville ou un département',
-        })
-      ).not.toHaveAttribute('disabled')
       expect(() =>
         within(etape2).getAllByRole('option', { hidden: true })
       ).toThrow()
     })
 
-    it('affiche le champ de saisie de l’offre si l’utilisateur souhaite faire une recherche par ID', async () => {
-      // Given
-      const etape2 = screen.getByRole('group', {
-        name: 'Étape 2: Critères de recherche',
-      })
-      const checkbox = screen.getByRole('checkbox', {
-        name: 'Recherche avec un numéro d’offre Pôle emploi',
-      })
-      // When
-      await userEvent.click(checkbox)
-
-      // Then
-      expect(within(etape2).getByLabelText(/Numéro d’offre/)).toHaveAttribute(
-        'type',
-        'text'
-      )
-      expect(
-        within(etape2).getByLabelText('Mots clés (Métier, code ROME)')
-      ).toHaveAttribute('disabled')
-      expect(
-        within(etape2).getByRole('combobox', {
-          name: 'Lieu de travail Saisissez une ville ou un département',
-        })
-      ).toHaveAttribute('disabled')
-    })
-
     describe('autocomplétion localisation', () => {
       it('retire les accents à la saisie', async () => {
         // Given
-        const inputLocalisation = screen.getByLabelText(/Lieu de travail/)
+        const inputLocalisation = screen.getByLabelText(/Localisation/)
 
         // When
-        await saisirLocalite('Ardèche')
+        await saisirCommune('Épinay-sur-Seine')
 
         // Then
-        expect(inputLocalisation).toHaveValue('ARDECHE')
-        expect(getCommunesEtDepartements).toHaveBeenCalledTimes(1)
-        expect(getCommunesEtDepartements).toHaveBeenCalledWith('ARDECHE')
+        expect(inputLocalisation).toHaveValue('EPINAY SUR SEINE')
+        expect(getCommunes).toHaveBeenCalledTimes(1)
+        expect(getCommunes).toHaveBeenCalledWith('EPINAY SUR SEINE')
       })
 
-      it('récupère les communes et les départements à la saisie', async () => {
+      it('récupère les communes à la saisie', async () => {
         // When
-        await saisirLocalite('Paris')
+        await saisirCommune('Epinay-sur-Seine')
 
         // Then
-        expect(getCommunesEtDepartements).toHaveBeenCalledTimes(1)
-        expect(getCommunesEtDepartements).toHaveBeenCalledWith('PARIS')
+        expect(getCommunes).toHaveBeenCalledTimes(1)
+        expect(getCommunes).toHaveBeenCalledWith('EPINAY SUR SEINE')
         expect(screen.getAllByRole('option', { hidden: true })).toHaveLength(
-          localites.length
+          communes.length
         )
-        localites.forEach((localite) => {
+        communes.forEach((commune) => {
           expect(
             screen.getByRole('option', {
               hidden: true,
-              name: sanitize(localite.libelle),
+              name: commune.libelle,
             })
-          ).toHaveValue(sanitize(localite.libelle))
+          ).toHaveValue(commune.libelle)
         })
       })
 
       it('affiche une erreur quand la localisation est incorrecte', async () => {
         // Given
-        const inputLocalisation = screen.getByLabelText(/Lieu de travail/)
+        const inputLocalisation = screen.getByLabelText(/Localisation/)
         const submitButton = screen.getByRole('button', {
           name: 'Rechercher',
         })
@@ -162,9 +125,9 @@ describe('Page Recherche Alternances', () => {
 
         // Then
         expect(
-          screen.getByText('Veuillez saisir une localisation correcte.')
+          screen.getByText('Veuillez saisir une commune correcte.')
         ).toBeInTheDocument()
-        expect(searchAlternances).toHaveBeenCalledTimes(0)
+        expect(searchServicesCiviques).toHaveBeenCalledTimes(0)
       })
     })
 
@@ -182,14 +145,10 @@ describe('Page Recherche Alternances', () => {
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
         // Then
-        const etape3 = screen.getByRole('group', {
-          name: 'Étape 3: Plus de critères',
-        })
-        expect(etape3).toBeInTheDocument()
         expect(screen.getByText('Voir moins de critères')).toBeInTheDocument()
       })
 
-      it('ne permet pas de selectionner de type de contrat', async () => {
+      it('permet de selectionner un domaine', async () => {
         // When
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
@@ -197,14 +156,19 @@ describe('Page Recherche Alternances', () => {
         const etape3 = screen.getByRole('group', {
           name: 'Étape 3: Plus de critères',
         })
-        expect(() =>
-          within(etape3).getByRole('group', {
-            name: 'Type de contrat',
-          })
-        ).toThrow()
+        const selectDomaine = within(etape3).getByRole('combobox', {
+          name: 'Sélectionner domaine',
+        })
+        domainesServiceCivique.forEach((domaine) => {
+          expect(
+            within(selectDomaine).getByRole('option', {
+              name: domaine.libelle,
+            })
+          ).toHaveValue(domaine.code)
+        })
       })
 
-      it('permet de selectionner un ou plusieurs temps de travail', async () => {
+      it("permet d'afficher uniquement les offres débutant dès que possible", async () => {
         // When
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
@@ -212,38 +176,59 @@ describe('Page Recherche Alternances', () => {
         const etape3 = screen.getByRole('group', {
           name: 'Étape 3: Plus de critères',
         })
-        const tempsTravailGroup = within(etape3).getByRole('group', {
-          name: 'Temps de travail',
+        const dateDebutGroupe = within(etape3).getByRole('group', {
+          name: 'Date de début',
         })
-        expect(tempsTravailGroup).toBeInTheDocument()
         expect(
-          within(tempsTravailGroup).getByRole('checkbox', {
-            name: 'Temps partiel',
+          within(dateDebutGroupe).getByRole('checkbox', {
+            name: /Dès que possible/,
           })
-        ).toBeInTheDocument()
+        ).toBeChecked()
         expect(
-          within(tempsTravailGroup).getByRole('checkbox', {
-            name: 'Temps plein',
-          })
-        ).toBeInTheDocument()
-      })
-
-      it("permet d'afficher uniquement les offres débutant accepté", async () => {
-        // When
-        await userEvent.click(screen.getByText('Voir plus de critères'))
-
-        // Then
-        const etape3 = screen.getByRole('group', {
-          name: 'Étape 3: Plus de critères',
-        })
-        const experienceGroup = within(etape3).getByRole('group', {
-          name: 'Expérience',
-        })
-        expect(
-          within(experienceGroup).getByRole('checkbox', {
-            name: /Afficher uniquement les offres débutant accepté/,
+          within(dateDebutGroupe).getByRole('checkbox', {
+            name: /À partir de/,
           })
         ).not.toBeChecked()
+        expect(
+          within(dateDebutGroupe).getByRole('checkbox', {
+            name: /À partir de/,
+          })
+        ).toHaveAttribute('disabled')
+      })
+
+      it('permet de définir une date de début', async () => {
+        // Given
+        const today = '2022-10-12'
+        jest.spyOn(DateTime, 'now').mockReturnValue(DateTime.fromISO(today))
+
+        // When
+        await userEvent.click(screen.getByText('Voir plus de critères'))
+        const etape3 = screen.getByRole('group', {
+          name: 'Étape 3: Plus de critères',
+        })
+        expect(() => within(etape3).getByLabelText('Date de début')).toThrow()
+
+        await userEvent.click(screen.getByLabelText(/Dès que possible/))
+
+        // Then
+        const dateDebutGroupe = within(etape3).getByRole('group', {
+          name: 'Date de début',
+        })
+        expect(
+          within(dateDebutGroupe).getByRole('checkbox', {
+            name: /Dès que possible/,
+          })
+        ).not.toBeChecked()
+        expect(
+          within(dateDebutGroupe).getByRole('checkbox', {
+            name: /À partir de/,
+          })
+        ).toBeChecked()
+        const inputDate = within(dateDebutGroupe).getByLabelText(
+          'Sélectionner une date de début'
+        )
+        expect(inputDate).toHaveAttribute('type', 'date')
+        expect(inputDate).toHaveAttribute('value', today)
       })
 
       it('permet de définir un rayon de recherche si une commune est sélectionnée', async () => {
@@ -256,7 +241,7 @@ describe('Page Recherche Alternances', () => {
           within(etape3).getByRole('group', { name: 'Distance' })
         ).toThrow()
 
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
 
         // Then
         const distanceGroup = within(etape3).getByRole('group', {
@@ -276,17 +261,20 @@ describe('Page Recherche Alternances', () => {
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
         // When-Then
-        await userEvent.click(
-          screen.getByLabelText(
-            /Afficher uniquement les offres débutant accepté/
-          )
+        await userEvent.selectOptions(
+          screen.getByLabelText('Sélectionner domaine'),
+          domainesServiceCivique[2].libelle
         )
         expect(screen.getByText('[1] filtre sélectionné')).toBeInTheDocument()
 
-        await userEvent.click(screen.getByLabelText('Temps plein'))
+        await userEvent.click(screen.getByLabelText(/Dès que possible/))
+        fireEvent.change(
+          screen.getByLabelText('Sélectionner une date de début'),
+          { target: { value: '2022-11-01' } }
+        )
         expect(screen.getByText('[2] filtres sélectionnés')).toBeInTheDocument()
 
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
         fireEvent.change(screen.getByLabelText(/Dans un rayon de/), {
           target: { value: 43 },
         })
@@ -298,14 +286,18 @@ describe('Page Recherche Alternances', () => {
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
         // Then
-        expect(screen.getByLabelText(/débutant accepté/)).toBeChecked()
-        expect(screen.getByLabelText(/Temps plein/)).toBeChecked()
+        expect(screen.getByLabelText(/domaine/)).toHaveValue(
+          domainesServiceCivique[2].code
+        )
+        expect(screen.getByLabelText(/Dès que possible/)).not.toBeChecked()
+        expect(screen.getByLabelText(/À partir de/)).toBeChecked()
+        expect(screen.getByLabelText(/date de début/)).toHaveValue('2022-11-01')
         expect(screen.getByLabelText(/rayon/)).toHaveValue('43')
       })
     })
 
     describe('partage des critères de recherche', () => {
-      it('ne permet pas de partager s’il n’y a ni mots clés ni localité renseignés', async () => {
+      it('ne permet pas de partager s’il n’y a pas de commune renseignée', async () => {
         await userEvent.click(
           screen.getByRole('button', {
             name: `Partager les critères de recherche`,
@@ -314,16 +306,14 @@ describe('Page Recherche Alternances', () => {
 
         expect(
           screen.getByText(
-            'Pour suggérer des critères de recherche, vous devez saisir un mot clé et un lieu de travail.'
+            'Pour suggérer des critères de recherche, vous devez saisir une ville.'
           )
         ).toBeInTheDocument()
       })
 
-      it('affiche le bouton de partage de critère s’il y a des mots clés et une localité renseignés', async () => {
+      it('affiche le bouton de partage de critère s’il y a une commune renseignés', async () => {
         // When
-        const inputMotsCles = screen.getByLabelText(/Mots clés/)
-        await userEvent.type(inputMotsCles, 'Prof')
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
 
         // Then
         expect(
@@ -340,9 +330,7 @@ describe('Page Recherche Alternances', () => {
 
       it('construit le bon lien qui correspond aux critères de recherches', async () => {
         // Given
-        const inputMotsCles = screen.getByLabelText(/Mots clés/)
-        await userEvent.type(inputMotsCles, 'Prof')
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
 
         // Then
         expect(
@@ -351,13 +339,13 @@ describe('Page Recherche Alternances', () => {
           })
         ).toHaveAttribute(
           'href',
-          `/offres/partage-recherche?type=ALTERNANCE&titre=Prof%20-%20PARIS%2014%20(75)&motsCles=Prof&typeLocalite=COMMUNE&labelLocalite=PARIS%2014%20(75)&codeLocalite=75114`
+          `/offres/partage-recherche?type=SERVICE_CIVIQUE&titre=PARIS%2014%20(75)&labelLocalite=PARIS%2014%20(75)&latitude=48.830108&longitude=2.323026`
         )
       })
     })
 
     describe('recherche', () => {
-      it('permet de rechercher des alternances', async () => {
+      it('permet de rechercher des offres de services civiques', async () => {
         // Given
         const submitButton = screen.getByRole('button', {
           name: 'Rechercher',
@@ -367,69 +355,22 @@ describe('Page Recherche Alternances', () => {
         await userEvent.click(submitButton)
 
         // Then
-        expect(searchAlternances).toHaveBeenCalledWith({}, 1)
-      })
-
-      it('permet de faire ue recherche d’offre par Id ', async () => {
-        // Given
-        const checkbox = screen.getByRole('checkbox', {
-          name: 'Recherche avec un numéro d’offre Pôle emploi',
-        })
-        await userEvent.click(checkbox)
-        await userEvent.type(
-          screen.getByLabelText(/Numéro d’offre/),
-          'id-offre'
-        )
-
-        const submitButton = screen.getByRole('button', {
-          name: 'Rechercher',
-        })
-
-        // When
-        await userEvent.click(submitButton)
-
-        // Then
-        expect(getOffreEmploiClientSide).toHaveBeenCalledWith('id-offre')
-      })
-
-      it('construit la recherche avec un département', async () => {
-        // Given
-        const inputMotsCles = screen.getByLabelText(/Mots clés/)
-        const submitButton = screen.getByRole('button', {
-          name: 'Rechercher',
-        })
-
-        // When
-        await userEvent.type(inputMotsCles, 'prof industrie')
-        await saisirLocalite('pAris')
-        await userEvent.click(submitButton)
-
-        // Then
-        expect(searchAlternances).toHaveBeenCalledWith(
-          {
-            motsCles: 'prof industrie',
-            departement: unDepartement(),
-          },
-          1
-        )
+        expect(searchServicesCiviques).toHaveBeenCalledWith({}, 1)
       })
 
       it('construit la recherche avec une commune', async () => {
         // Given
-        const inputMotsCles = screen.getByLabelText(/Mots clés/)
         const submitButton = screen.getByRole('button', {
           name: 'Rechercher',
         })
 
         // When
-        await userEvent.type(inputMotsCles, 'prof industrie')
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
         await userEvent.click(submitButton)
 
         // Then
-        expect(searchAlternances).toHaveBeenCalledWith(
+        expect(searchServicesCiviques).toHaveBeenCalledWith(
           {
-            motsCles: 'prof industrie',
             commune: uneCommune(),
             rayon: 10,
           },
@@ -439,19 +380,18 @@ describe('Page Recherche Alternances', () => {
 
       it('construit la recherche avec les critères d’affinage', async () => {
         // Given
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
-        await userEvent.click(
-          screen.getByLabelText(
-            /Afficher uniquement les offres débutant accepté/
-          )
+        await userEvent.selectOptions(
+          screen.getByLabelText('Sélectionner domaine'),
+          domainesServiceCivique[2].libelle
         )
-
-        await userEvent.click(screen.getByLabelText('Temps plein'))
-        await userEvent.click(screen.getByLabelText('Temps partiel'))
-        await userEvent.click(screen.getByLabelText('Temps partiel'))
-
+        await userEvent.click(screen.getByLabelText(/Dès que possible/))
+        fireEvent.change(
+          screen.getByLabelText('Sélectionner une date de début'),
+          { target: { value: '2022-11-01' } }
+        )
         fireEvent.change(screen.getByLabelText(/Dans un rayon de/), {
           target: { value: 43 },
         })
@@ -462,11 +402,11 @@ describe('Page Recherche Alternances', () => {
         )
 
         // Then
-        expect(searchAlternances).toHaveBeenCalledWith(
+        expect(searchServicesCiviques).toHaveBeenCalledWith(
           {
             commune: uneCommune(),
-            debutantAccepte: true,
-            durees: ['Temps plein'],
+            domaine: domainesServiceCivique[2].code,
+            dateDebut: '2022-11-01',
             rayon: 43,
           },
           1
@@ -475,22 +415,24 @@ describe('Page Recherche Alternances', () => {
 
       it('vide les critères lorsqu’on change le type d’offre', async () => {
         // Given
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
         await userEvent.click(screen.getByText('Voir plus de critères'))
 
-        await userEvent.click(
-          screen.getByLabelText(
-            /Afficher uniquement les offres débutant accepté/
-          )
+        await userEvent.selectOptions(
+          screen.getByLabelText('Sélectionner domaine'),
+          domainesServiceCivique[2].libelle
         )
-
-        await userEvent.click(screen.getByLabelText('Temps plein'))
+        await userEvent.click(screen.getByLabelText(/Dès que possible/))
+        fireEvent.change(
+          screen.getByLabelText('Sélectionner une date de début'),
+          { target: { value: '2022-11-01' } }
+        )
         fireEvent.change(screen.getByLabelText(/Dans un rayon de/), {
           target: { value: 43 },
         })
 
+        await userEvent.click(screen.getByText('Immersion'))
         await userEvent.click(screen.getByText('Service civique'))
-        await userEvent.click(screen.getByText('Alternance'))
 
         // When
         await userEvent.click(
@@ -498,7 +440,7 @@ describe('Page Recherche Alternances', () => {
         )
 
         // Then
-        expect(searchAlternances).toHaveBeenCalledWith({}, 1)
+        expect(searchServicesCiviques).toHaveBeenCalledWith({}, 1)
       })
     })
 
@@ -515,53 +457,45 @@ describe('Page Recherche Alternances', () => {
 
         // Then
         offresList = screen.getByRole('list', {
-          description: `Liste des résultats (37 offres)`,
+          description: 'Liste des résultats (37 offres)',
         })
       })
 
-      it('affiche les offres', async () => {
+      it('affiche toutes les offres', async () => {
         expect(within(offresList).getAllByRole('listitem').length).toEqual(
-          alternances.length
+          servicesCiviques.length
         )
       })
 
       it('affiche chaque offre', async () => {
-        alternances.forEach((alternance) => {
+        servicesCiviques.forEach((offre) => {
           const offreCard = within(offresList).getByRole('heading', {
             level: 3,
-            name: 'Offre n°' + alternance.id,
+            name: offre.titre,
           }).parentElement!
-          expect(within(offreCard).getByText('Alternance')).toBeInTheDocument()
+          expect(within(offreCard).getByText(offre.domaine)).toBeInTheDocument()
           expect(
-            within(offreCard).getByText(alternance.titre)
+            within(offreCard).getByText(offre.organisation!)
           ).toBeInTheDocument()
-          expect(
-            within(offreCard).getByText(alternance.typeContrat)
-          ).toBeInTheDocument()
-          expect(
-            within(offreCard).getByText(alternance.duree!)
-          ).toBeInTheDocument()
-          expect(
-            within(offreCard).getByText(alternance.nomEntreprise!)
-          ).toBeInTheDocument()
-          expect(
-            within(offreCard).getByText(alternance.localisation!)
-          ).toBeInTheDocument()
+          expect(within(offreCard).getByText(offre.ville!)).toBeInTheDocument()
           expect(
             within(offreCard).getByRole('link', {
-              name: 'Voir le détail de l’offre ' + alternance.titre,
+              name: 'Voir le détail de l’offre ' + offre.titre,
             })
-          ).toHaveAttribute('href', '/offres/emploi/' + alternance.id)
+          ).toHaveAttribute('href', '/offres/service-civique/' + offre.id)
         })
       })
 
       it('permet de partager chaque offre', () => {
-        alternances.forEach((offre) => {
+        servicesCiviques.forEach((offre) => {
           expect(
             within(offresList).getByRole('link', {
               name: `Partager offre ${offre.titre}`,
             })
-          ).toHaveAttribute('href', `/offres/emploi/${offre.id}/partage`)
+          ).toHaveAttribute(
+            'href',
+            `/offres/service-civique/${offre.id}/partage`
+          )
         })
       })
 
@@ -590,7 +524,7 @@ describe('Page Recherche Alternances', () => {
         expect(within(offresList).getAllByRole('listitem').length).toBeTruthy()
 
         // When
-        await userEvent.type(screen.getByLabelText(/Mots clés/), 'Boulanger')
+        await userEvent.type(screen.getByLabelText(/Localisation/), 'Rennes')
 
         // Then
         expect(
@@ -602,10 +536,10 @@ describe('Page Recherche Alternances', () => {
 
       describe('pagination', () => {
         beforeEach(() => {
-          ;(searchAlternances as jest.Mock).mockImplementation(
+          ;(searchServicesCiviques as jest.Mock).mockImplementation(
             (_query, page) => ({
               metadonnees: { nombreTotal: 37, nombrePages: 4 },
-              offres: [uneBaseAlternance({ titre: 'Offre page ' + page })],
+              offres: [uneBaseServiceCivique({ titre: 'Offre page ' + page })],
             })
           )
         })
@@ -615,7 +549,7 @@ describe('Page Recherche Alternances', () => {
           await userEvent.click(screen.getByLabelText('Page 2'))
 
           // Then
-          expect(searchAlternances).toHaveBeenCalledWith({}, 2)
+          expect(searchServicesCiviques).toHaveBeenCalledWith({}, 2)
           expect(screen.getByText('Offre page 2')).toBeInTheDocument()
         })
 
@@ -625,8 +559,8 @@ describe('Page Recherche Alternances', () => {
           await userEvent.click(screen.getByLabelText('Page suivante'))
 
           // Then
-          expect(searchAlternances).toHaveBeenCalledWith({}, 2)
-          expect(searchAlternances).toHaveBeenCalledWith({}, 3)
+          expect(searchServicesCiviques).toHaveBeenCalledWith({}, 2)
+          expect(searchServicesCiviques).toHaveBeenCalledWith({}, 3)
 
           expect(screen.getByLabelText(`Page 3`)).toHaveAttribute(
             'aria-current',
@@ -639,7 +573,7 @@ describe('Page Recherche Alternances', () => {
           await userEvent.click(screen.getByLabelText(`Page 1`))
 
           // Then
-          expect(searchAlternances).toHaveBeenCalledTimes(1)
+          expect(searchServicesCiviques).toHaveBeenCalledTimes(1)
         })
       })
     })
@@ -647,48 +581,47 @@ describe('Page Recherche Alternances', () => {
     describe('sauvegarde', () => {
       it('retient l’état de la recherche', async () => {
         // Given
-        await userEvent.type(
-          screen.getByLabelText(/Mots clés/),
-          'prof industrie'
-        )
-        await saisirLocalite('paris 14 (75)')
+        await saisirCommune('paris 14 (75)')
         await userEvent.click(screen.getByText('Voir plus de critères'))
-
-        await userEvent.click(
-          screen.getByLabelText(
-            /Afficher uniquement les offres débutant accepté/
-          )
+        await userEvent.selectOptions(
+          screen.getByLabelText('Sélectionner domaine'),
+          domainesServiceCivique[2].libelle
         )
-        await userEvent.click(screen.getByLabelText('Temps plein'))
+        await userEvent.click(screen.getByLabelText(/Dès que possible/))
+        fireEvent.change(
+          screen.getByLabelText('Sélectionner une date de début'),
+          { target: { value: '2022-11-01' } }
+        )
         fireEvent.change(screen.getByLabelText(/Dans un rayon de/), {
           target: { value: 43 },
         })
         await userEvent.click(
-          screen.getByRole('button', {
-            name: 'Rechercher',
-          })
+          screen.getByRole('button', { name: 'Rechercher' })
         )
         await userEvent.click(screen.getByRole('button', { name: 'Page 2' }))
 
         // When
         rendered.unmount()
-        renderWithContexts(<RechercheOffres pageTitle='' />, {})
+        renderWithContexts(<RechercheOffresPage />, {})
 
         // Then
-        expect(screen.getByLabelText('Alternance')).toBeChecked()
-        expect(screen.getByLabelText(/Lieu de travail/)).toHaveValue(
+        expect(screen.getByLabelText('Service civique')).toBeChecked()
+        expect(screen.getByLabelText(/Localisation/)).toHaveValue(
           'PARIS 14 (75)'
         )
         expect(screen.getByText('[3] filtres sélectionnés')).toBeInTheDocument()
         await userEvent.click(screen.getByText('Voir plus de critères'))
-        expect(screen.getByLabelText(/débutant accepté/)).toBeChecked()
-        expect(screen.getByLabelText('Temps plein')).toBeChecked()
+        expect(screen.getByLabelText(/domaine/)).toHaveValue(
+          domainesServiceCivique[2].code
+        )
+        expect(screen.getByLabelText(/Dès que possible/)).not.toBeChecked()
+        expect(screen.getByLabelText(/date de début/)).toHaveValue('2022-11-01')
         expect(screen.getByLabelText(/rayon/)).toHaveValue('43')
         const offresList = screen.getByRole('list', {
           description: 'Liste des résultats (37 offres)',
         })
         expect(within(offresList).getAllByRole('listitem').length).toEqual(
-          alternances.length
+          servicesCiviques.length
         )
         expect(screen.getByRole('button', { name: 'Page 2' })).toHaveAttribute(
           'aria-current',
@@ -697,28 +630,23 @@ describe('Page Recherche Alternances', () => {
       })
     })
   })
+
   describe('quand le conseiller est PE BRSA', () => {
     it('n’affiche pas la recherche en tant que conseiller PE BRSA', () => {
-      rendered = renderWithContexts(<RechercheOffres pageTitle='' />, {
+      rendered = renderWithContexts(<RechercheOffresPage />, {
         customConseiller: unConseiller({
           structure: StructureConseiller.POLE_EMPLOI_BRSA,
         }),
       })
       // Then
-      expect(() => screen.getByRole('radio', { name: 'Alternance' })).toThrow()
+      expect(() =>
+        screen.getByRole('radio', { name: 'Service civique' })
+      ).toThrow()
     })
   })
 })
 
-async function saisirLocalite(text: string) {
-  await userEvent.type(screen.getByLabelText(/Lieu de travail/), text)
+async function saisirCommune(text: string) {
+  await userEvent.type(screen.getByLabelText(/Localisation/), text)
   await act(() => new Promise((r) => setTimeout(r, 500)))
-}
-
-function sanitize(str: string): string {
-  return str
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .replace(/['-]/g, ' ')
-    .toUpperCase()
 }
